+++ packages/plugin-commons/web/src/Icons/RTLIcon.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 import React from 'react';
 
-export const RTLIcon = props => (
+const RTLIcon = props => (
   <svg width="24px" height="26px" viewBox="0 0 24 26" {...props}>
     <g stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
       <g transform="translate(-152.000000, -583.000000)">
@@ -33,3 +33,5 @@ export const RTLIcon = props => (
     </g>
   </svg>
 );
+
+export default RTLIcon;