+++ packages/plugin-social-polls/web/src/assets/icons/index.ts
@@ -17,6 +17,3 @@ export * from './GradientIcon';
 
 export * from './LayoutGridIcon';
 export * from './LayoutListIcon';
-
-export * from './RTLIcon';
-export * from './LTRIcon';