+++ packages/plugin-social-polls/web/src/components/settings/layout-settings-section/LayoutSettingsSection.jsx
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import { SelectionList, LabeledToggle } from 'wix-rich-content-plugin-commons';
+import { SelectionList, LabeledToggle, LTRIcon, RTLIcon } from 'wix-rich-content-plugin-commons';
 import { Separator, InfoIcon } from 'wix-rich-content-editor-common';
 import { mergeStyles } from 'wix-rich-content-common';
 
 import { LAYOUT, DIRECTION } from '../../../defaults';
-import { LayoutGridIcon, LayoutListIcon, LTRIcon, RTLIcon } from '../../../assets/icons';
+import { LayoutGridIcon, LayoutListIcon } from '../../../assets/icons';
 
 import styles from './layout-settings-section.scss';