+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 ### 🏠 Internal
 - `editor`
   - [#1556](https://github.com/wix-incubator/rich-content/pull/1556) fix editorStateConversion import size
+- `plugin-commons`
+  - [#1566](https://github.com/wix-incubator/rich-content/pull/1566) move rtl/ltr icons froms polls to plugin-commons package
 
   
   </details>